container-setup.bash: Additional configuration for container
The container does not have installed the Python requirements needed
to run some of the scripts under /opt/proof-generation. This installs
them (using the requirements.txt in that directory). It also sets up
a PYTHONPATH so you can use the scripts regardless of the CWD, and
adds a `psym` alias to run the prove_symbolic.py script.