sample/fast-exp-spec.k: Fix import name to match typo in container file
This `requires "verification.h"`, which we can make work by adding `-I
/opt/proof-generation-evaluation/definitions` to the `kompile` command
line, but in that required file the module is actually mis-spelled as
VERIFICAITON (swapped I and T near the end). Make fast-exp-spec.k match
this.

The module still doesn't kompile for other reasons; this

    Inner Parser: Scanner error: unexpected character sequence '!'.

is produced by `while (!(k <= 1))` in the claim.